+++ src/hfr.c
@@ -33,5 +33,6 @@
 #include "games/th12.c"
 #include "games/th13.c"
 #include "games/th14.c"
+#include "games/th15.c"
 #include "core/install.c"
 #include "core/entry.c"